RunIt/requirements.txt
aiodns==2.0.0
asyncio==3.4.3
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
fastapi==0.63.0
future==0.18.2
futures==3.1.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.10.1
httpcore==0.9.1
httptools==0.1.2
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
loguru==0.5.1
motor==2.2.0
pycares==4.0.0
pycparser==2.20
pydantic==1.6.1
PyJWT==2.1.0
pymongo==3.11.0
PyYAML==6.0
queuelib==1.5.0
redis==3.5.3
requests==2.24.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.13.6
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
Werkzeug==1.0.1
